Discarded mixin and placing AggregateField in HoverField
