+++ src/js/components/pipelines/AggregateField.jsx
@@ -1,24 +1,17 @@
 'use strict';
 
-var React = require('react'),
-    DraggableFields = require('../mixins/DraggableFields');
+var React = require('react');
 
 var AggregateField = React.createClass({
-  propTypes: {
-    dsId: React.PropTypes.number,
-    className: React.PropTypes.string.isRequired,
-    def: React.PropTypes.object,
-    schema: React.PropTypes.object
-  },
-  mixins: [DraggableFields],
   render: function() {
     var props = this.props;
     return (
       <div draggable={true}
-        onDragStart={this.handleDragStart}
-        onDragOver={this.handleDragOver}
-        onDragEnd={this.handleDragEnd}
-        className={'aggregate-field ' + props.className}>
+        onDragStart={props.onDragStart}
+        onDragOver={props.onDragOver}
+        onDragEnd={props.onDragEnd}
+        onDrop={props.onDrop}
+        className="aggregate-field">
         {props.aggregationName} - {props.field}
       </div>
     );
