+++ src/js/components/pipelines/HoverField.jsx
@@ -1,15 +1,17 @@
 'use strict';
+
 var React = require('react'),
     connect = require('react-redux').connect,
     dsUtil = require('../../util/dataset-utils'),
     bindChannel = require('../../actions/bindChannel'),
     FieldType = require('./FieldType'),
     SortField = require('./SortField'),
+    sg = require('../../ctrl/signals'),
+    ctrl = require('../../ctrl'),
     Icon = require('../Icon'),
     AggregateField = require('./AggregateField'),
     assets = require('../../util/assets'),
-    AGGREGATION_OPS = require('../../constants/transformTypes').aggregationOps,
-    DraggableFields = require('../mixins/DraggableFields');
+    AGGREGATION_OPS = require('../../constants/transformTypes').aggregationOps;
 
 var LIST_LIMIT = 5;
 
@@ -33,7 +35,7 @@ var HoverField = React.createClass({
     schema: React.PropTypes.object
   },
 
-  mixins: [DraggableFields],
+  // mixins: [DraggableFields],
 
   getInitialState: function() {
     return {
@@ -66,6 +68,69 @@ var HoverField = React.createClass({
     this.setState(state);
   },
 
+  handleDragStart: function(evt) {
+    var state = {
+      bindField: this.state.fieldDef
+    };
+
+    // if an AggregateField isn't being dragged, close the menu
+    // and reset listLimit
+    if (!evt.target.classList.contains('aggregate-field')) {
+      state.showFieldTransforms = false;
+      state.listLimit = LIST_LIMIT;
+    }
+
+    this.setState(state);
+    evt.dataTransfer.setData('text/plain', evt.target.id);
+    evt.dataTransfer.effectAllowed = 'link';
+    sg.set(sg.MODE, 'channels');
+    ctrl.update();
+  },
+
+  handleDragOver: function(evt) {
+    if (evt.preventDefault) {
+      evt.preventDefault(); // Necessary. Allows us to drop.
+    }
+
+    return false;
+  },
+
+  // This makes use of the bubble cursor, which corresponds to the cell signal;
+  // we're using that to figure out which channel we are closest to. The
+  // SELECTED signal indicates the mark to bind the data to.
+  handleDragEnd: function(evt) {
+    var props = this.props,
+        sel = sg.get(sg.SELECTED),
+        cell = sg.get(sg.CELL),
+        bindField = this.state.bindField,
+        dropped = sel._id && cell._id;
+
+    try {
+      if (dropped) {
+        props.bindChannel(props.dsId, bindField, sel.mark.def.lyra_id, cell.key);
+      }
+    } catch (e) {
+      console.warn('Unable to bind primitive');
+      console.warn(e);
+    }
+
+    sg.set(sg.MODE, 'handles');
+    sg.set(sg.CELL, {});
+    this.setState({bindField: null});
+
+    if (!dropped) {
+      ctrl.update();
+    }
+  },
+
+  handleDrop: function(evt) {
+    if (evt.preventDefault) {
+      evt.preventDefault(); // Necessary. Allows us to drop.
+    }
+
+    return false;
+  },
+
   toggleTransforms: function(evt) {
     this.setState({
       showFieldTransforms: !this.state.showFieldTransforms
@@ -88,7 +153,13 @@ var HoverField = React.createClass({
     var state = this.state,
         field = state.fieldDef,
         style = {top: state.offsetTop, display: field ? 'block' : 'none'},
-        transforms = AGGREGATION_OPS.slice(0, state.listLimit);
+        transforms = AGGREGATION_OPS.slice(0, state.listLimit),
+        aggregateFieldProps = {
+          onDragStart: this.handleDragStart,
+          onDragOver: this.handleDragOver,
+          onDragEnd: this.handleDragEnd,
+          onDrop: this.handleDrop
+        };
 
     // Icon use temporary asset
     var transformsIcon = (<Icon onClick={this.toggleTransforms} glyph={assets.symbol}
@@ -113,7 +184,7 @@ var HoverField = React.createClass({
                     <li className="transform-item" key={i}>
                       <AggregateField aggregationName={transform}
                         field={field.name}
-                        className={this.props.className} />
+                        {...aggregateFieldProps} />
                     </li>
                   );
                 }, this)
